Move specific quests back to questcommand
