+++ src/lib/diaries.ts
@@ -2,6 +2,7 @@ import { objectEntries } from 'e';
 import { Monsters } from 'oldschooljs';
 import { Item } from 'oldschooljs/dist/meta/types';
 
+import { MAX_QP } from '../mahoji/lib/abstracted_commands/questCommand';
 import type { MinigameName, MinigameScore } from './settings/minigames';
 import Skillcapes from './skilling/skillcapes';
 import { courses } from './skilling/skills/agility';
@@ -10,7 +11,6 @@ import { Skills } from './types';
 import { formatSkillRequirements, hasSkillReqs, itemNameFromID } from './util';
 import getOSItem from './util/getOSItem';
 import resolveItems from './util/resolveItems';
-import { MAX_QP } from './util/specificQuests';
 
 export const diaryTiers = ['easy', 'medium', 'hard', 'elite'] as const;
 export type DiaryTierName = (typeof diaryTiers)[number];
