+++ src/lib/mastery.ts
@@ -1,12 +1,12 @@
 import { calcWhatPercent, clamp, round, sumArr } from 'e';
 
 import { calculateAchievementDiaryProgress } from '../mahoji/lib/abstracted_commands/achievementDiaryCommand';
+import { MAX_QP } from '../mahoji/lib/abstracted_commands/questCommand';
 import { allCombatAchievementTasks } from './combat_achievements/combatAchievements';
 import { MAX_XP } from './constants';
 import { getTotalCl } from './data/Collections';
 import { SkillsEnum } from './skilling/types';
 import { MUserStats } from './structures/MUserStats';
-import { MAX_QP } from './util/specificQuests';
 
 export async function calculateMastery(user: MUser, stats: MUserStats) {
 	const [totalClItems, clItems] = getTotalCl(user, 'collection', stats);
