+++ src/lib/minions/data/killableMonsters/bosses/dt.ts
@@ -1,13 +1,13 @@
 import { Time } from 'e';
 import { Bank, Monsters } from 'oldschooljs';
 
+import { QuestID } from '../../../../../mahoji/lib/abstracted_commands/questCommand';
 import { dukeSucellusCL, theLeviathanCL, theWhispererCL, vardorvisCL } from '../../../../data/CollectionsExport';
 import { GearStat } from '../../../../gear/types';
 import { SkillsEnum } from '../../../../skilling/types';
 import { roll } from '../../../../util';
 import itemID from '../../../../util/itemID';
 import resolveItems, { deepResolveItems } from '../../../../util/resolveItems';
-import { QuestID } from '../../../../util/specificQuests';
 import { KillableMonster } from '../../../types';
 
 const awakenedDeathProps = {
