+++ src/mahoji/commands/activities.ts
@@ -10,7 +10,6 @@ import birdhouses from '../../lib/skilling/skills/hunter/birdHouseTrapping';
 import { Castables } from '../../lib/skilling/skills/magic/castables';
 import { Enchantables } from '../../lib/skilling/skills/magic/enchantables';
 import Prayer from '../../lib/skilling/skills/prayer';
-import { quests } from '../../lib/util/specificQuests';
 import { aerialFishingCommand } from '../lib/abstracted_commands/aerialFishingCommand';
 import { alchCommand } from '../lib/abstracted_commands/alchCommand';
 import { birdhouseCheckCommand, birdhouseHarvestCommand } from '../lib/abstracted_commands/birdhousesCommand';
@@ -30,7 +29,7 @@ import { fightCavesCommand } from '../lib/abstracted_commands/fightCavesCommand'
 import { infernoStartCommand, infernoStatsCommand } from '../lib/abstracted_commands/infernoCommand';
 import { otherActivities, otherActivitiesCommand } from '../lib/abstracted_commands/otherActivitiesCommand';
 import puroOptions, { puroPuroStartCommand } from '../lib/abstracted_commands/puroPuroCommand';
-import { questCommand } from '../lib/abstracted_commands/questCommand';
+import { questCommand, quests } from '../lib/abstracted_commands/questCommand';
 import { sawmillCommand } from '../lib/abstracted_commands/sawmillCommand';
 import { scatterCommand } from '../lib/abstracted_commands/scatterCommand';
 import { underwaterAgilityThievingCommand } from '../lib/abstracted_commands/underwaterCommand';
