+++ src/tasks/minions/questingActivity.ts
@@ -5,7 +5,7 @@ import { SkillsEnum } from '../../lib/skilling/types';
 import type { ActivityTaskOptionsWithQuantity } from '../../lib/types/minions';
 import { roll } from '../../lib/util';
 import { handleTripFinish } from '../../lib/util/handleTripFinish';
-import { MAX_QP } from '../../lib/util/specificQuests';
+import { MAX_QP } from '../../mahoji/lib/abstracted_commands/questCommand';
 
 export const questingTask: MinionTask = {
 	type: 'Questing',
